SDK changes for workflow data errors
