+++ types/src/index.ts
@@ -180,8 +180,9 @@ export * from './dto/shipping/GetShipengineShippingResponse';
 export * from './dto/shipping/GetShippingLabelLegacyResponse';
 export * from './dto/shipping/SaveTrackingInfoParams';
 export * from './dto/shipping/ValidateAddressBody';
-export * from './dto/step-engine/Attachment';
-export * from './dto/step-engine/WorkflowStatus';
+export * from './dto/workflow/Attachment';
+export * from './dto/workflow/WorkflowError';
+export * from './dto/workflow/WorkflowStatus';
 export * from './dto/tag/UpsertTagParams';
 export * from './dto/taika/CreateCustomizationsTaikaParams';
 export * from './dto/taika/CustomizationOrderTaikaParams';
